Servicios Entidades creadas

+++ Backend/codequest/src/main/java/org/javeriana/codequest/repository/LearningMethodRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LearningMethodRepository extends JpaRepository<LearningMethod, Integer> {
+public interface LearningMethodRepository extends JpaRepository<LearningMethod, Long> {
 
 }
